+++ taplo-ide/vscode/src/server.ts
@@ -7,10 +7,16 @@
 
 import * as taplo from "taplo";
 
+import * as fs from "fs";
+
 (global as any).sendMessage = (msg: any) => {
   if (process.send) {
     process.send(msg);
   }
+};
+
+(global as any).readFile = (path: string): Uint8Array => {
+  return fs.readFileSync(path);
 };
 
 taplo.init();